Get job by Id.

+++ app/src/main/java/com/cupcake/jobsfinder/ui/create_job/state/JobFormUiState.kt
@@ -1,7 +1,7 @@
 package com.cupcake.jobsfinder.ui.create_job.state
 
 data class JobFormUiState(
-    val idJobTitle: Int = 0,
+    val idJobTitle: Long = 0,
     val company: String = "",
     val createdTime: String = "",
     val workType: String = "",
